requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
coverage==7.6.9
discord.py==2.4.0
Django==5.1.2
django-cors-headers==4.5.0
django-dotenv==1.4.2
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.1
Faker==33.1.0
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
mypy==1.11.1
mypy-extensions==1.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==73.0.1
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.44.0
whitenoise==6.7.0
yarl==1.9.4
mysqlclient==2.2.4